requirements.txt
appdirs==1.4.0
Django==1.10.5
django-filter==1.0.1
django-thumbs==0.4
djangorestframework==3.5.3
Markdown==2.6.8
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
six==1.10.0
